requirements.txt
numpy
pandas
requests
newscatcherapi
python-dotenv
yfinance
matplotlib
statsmodels
seaborn
# scikit-learn
nltk
# transformers
# kafka-python
pymongo
# mysql-connector-python
# redis-py
# streamlit
-e .